refactor temporal code
