+++ polars/Cargo.toml
@@ -153,8 +153,8 @@ dtype-datetime = [
   "polars-io/dtype-datetime",
   "polars-time/dtype-datetime",
 ]
-dtype-duration = ["polars-core/dtype-duration", "polars-lazy/dtype-duration"]
-dtype-time = ["polars-core/dtype-time", "polars-io/dtype-time"]
+dtype-duration = ["polars-core/dtype-duration", "polars-lazy/dtype-duration", "polars-time/dtype-duration"]
+dtype-time = ["polars-core/dtype-time", "polars-io/dtype-time", "polars-time/dtype-time"]
 dtype-i8 = ["polars-core/dtype-i8", "polars-lazy/dtype-i8"]
 dtype-i16 = ["polars-core/dtype-i16", "polars-lazy/dtype-i16"]
 dtype-u8 = ["polars-core/dtype-u8", "polars-lazy/dtype-u8"]
