+++ polars/polars-core/src/chunked_array/temporal/duration.rs
@@ -93,49 +93,4 @@ impl DurationChunked {
         let vals = v.into_iter().map(|opt| opt.map(func));
         Int64Chunked::from_iter_options(name, vals).into_duration(tu)
     }
-
-    /// Extract the hours from a `Duration`
-    pub fn hours(&self) -> Int64Chunked {
-        match self.time_unit() {
-            TimeUnit::Milliseconds => &self.0 / (MILLISECONDS * SECONDS_IN_HOUR),
-            TimeUnit::Microseconds => &self.0 / (MICROSECONDS * SECONDS_IN_HOUR),
-            TimeUnit::Nanoseconds => &self.0 / (NANOSECONDS * SECONDS_IN_HOUR),
-        }
-    }
-
-    /// Extract the days from a `Duration`
-    pub fn days(&self) -> Int64Chunked {
-        match self.time_unit() {
-            TimeUnit::Milliseconds => &self.0 / MILLISECONDS_IN_DAY,
-            TimeUnit::Microseconds => &self.0 / (MICROSECONDS * SECONDS_IN_DAY),
-            TimeUnit::Nanoseconds => &self.0 / (NANOSECONDS * SECONDS_IN_DAY),
-        }
-    }
-
-    /// Extract the milliseconds from a `Duration`
-    pub fn milliseconds(&self) -> Int64Chunked {
-        match self.time_unit() {
-            TimeUnit::Milliseconds => self.0.clone(),
-            TimeUnit::Microseconds => self.0.clone() / 1000,
-            TimeUnit::Nanoseconds => &self.0 / 1_000_000,
-        }
-    }
-
-    /// Extract the nanoseconds from a `Duration`
-    pub fn nanoseconds(&self) -> Int64Chunked {
-        match self.time_unit() {
-            TimeUnit::Milliseconds => &self.0 * NANOSECONDS_IN_MILLISECOND,
-            TimeUnit::Microseconds => &self.0 * 1000,
-            TimeUnit::Nanoseconds => self.0.clone(),
-        }
-    }
-
-    /// Extract the seconds from a `Duration`
-    pub fn seconds(&self) -> Int64Chunked {
-        match self.time_unit() {
-            TimeUnit::Milliseconds => &self.0 / MILLISECONDS,
-            TimeUnit::Microseconds => &self.0 / MICROSECONDS,
-            TimeUnit::Nanoseconds => &self.0 / NANOSECONDS,
-        }
-    }
 }
