+++ polars/polars-lazy/Cargo.toml
@@ -26,9 +26,9 @@ dtype-u8 = ["polars-core/dtype-u8"]
 dtype-u16 = ["polars-core/dtype-u16"]
 dtype-i8 = ["polars-core/dtype-i8"]
 dtype-i16 = ["polars-core/dtype-i16"]
-dtype-date = ["polars-core/dtype-date"]
-dtype-datetime = ["polars-core/dtype-datetime"]
-dtype-duration = ["polars-core/dtype-duration"]
+dtype-date = ["polars-core/dtype-date", "polars-time/dtype-date"]
+dtype-datetime = ["polars-core/dtype-datetime", "polars-time/dtype-datetime"]
+dtype-duration = ["polars-core/dtype-duration", "polars-time/dtype-duration"]
 dtype-categorical = ["polars-core/dtype-categorical"]
 dtype-struct = ["polars-core/dtype-struct"]
 object = ["polars-core/object"]
