+++ polars/polars-time/src/prelude.rs
@@ -1,3 +1,5 @@
 pub use crate::windows::{bounds::*, duration::*, groupby::*, window::*};
 pub use crate::*;
 pub use date_range::*;
+pub use crate::series::TemporalMethods;
+pub use crate::chunkedarray::*;
